update App added new features for redux
